// frontend/src/api/generated/zod/turmas/turmas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import * as zod from 'zod';


export const TurmasControllerCreateBody = zod.object({
  "semestreId": zod.uuid(),
  "cursoTecnico": zod.string(),
  "anoSerie": zod.string(),
  "turno": zod.enum(['MANHA', 'TARDE'])
})

export const TurmasControllerCreateResponse = zod.void()

export const TurmasControllerFindAllQueryParams = zod.object({
  "semestreId": zod.string().optional()
})

export const TurmasControllerFindAllResponseItem = zod.object({
  "id": zod.string(),
  "semestreId": zod.string(),
  "cursoTecnico": zod.string(),
  "anoSerie": zod.string(),
  "turno": zod.enum(['MANHA', 'TARDE']),
  "semestre": zod.object({
  "id": zod.string(),
  "nome": zod.string(),
  "dataInicio": zod.iso.datetime({"offset":true}),
  "dataFim": zod.iso.datetime({"offset":true})
})
})
export const TurmasControllerFindAllResponse = zod.array(TurmasControllerFindAllResponseItem)

export const TurmasControllerFindOneParams = zod.object({
  "id": zod.string()
})

export const TurmasControllerFindOneResponse = zod.unknown()

export const TurmasControllerUpdateParams = zod.object({
  "id": zod.string()
})

export const TurmasControllerUpdateBody = zod.object({
  "semestreId": zod.uuid().optional(),
  "cursoTecnico": zod.string().optional(),
  "anoSerie": zod.string().optional(),
  "turno": zod.enum(['MANHA', 'TARDE']).optional()
})

export const TurmasControllerUpdateResponse = zod.unknown()

export const TurmasControllerRemoveParams = zod.object({
  "id": zod.string()
})

export const TurmasControllerRemoveResponse = zod.unknown()