// frontend/src/api/generated/models/turmaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { SemestreDto } from './semestreDto';
import type { TurmaDtoTurno } from './turmaDtoTurno';

export interface TurmaDto {
  id: string;
  semestreId: string;
  cursoTecnico: string;
  anoSerie: string;
  turno: TurmaDtoTurno;
  semestre: SemestreDto;
}